Refactor general behavior RxViewModel into BaseViewModel class
