+++ app/src/main/java/com/marverenic/music/ui/RxViewModel.java
@@ -1,12 +1,8 @@
 package com.marverenic.music.ui;
 
-import android.content.Context;
-import android.content.res.Resources;
 import android.databinding.BaseObservable;
 import android.support.annotation.CheckResult;
-import android.support.annotation.DimenRes;
 import android.support.annotation.NonNull;
-import android.support.annotation.StringRes;
 
 import com.trello.rxlifecycle.FragmentEvent;
 import com.trello.rxlifecycle.LifecycleTransformer;
@@ -18,37 +14,15 @@ import rx.Observable;
 public abstract class RxViewModel extends BaseObservable {
 
     private Observable<FragmentEvent> mFragmentLifecycle;
-    private Context mContext;
 
     public RxViewModel(RxFragment fragment) {
         mFragmentLifecycle = fragment.lifecycle();
-        mContext = fragment.getContext();
     }
 
     @NonNull
     @CheckResult
     protected final <T> LifecycleTransformer<T> bindToLifecycle() {
         return RxLifecycle.bindUntilEvent(mFragmentLifecycle, FragmentEvent.DESTROY_VIEW);
-    }
-
-    protected Context getContext() {
-        return mContext;
-    }
-
-    protected Resources getResources() {
-        return mContext.getResources();
-    }
-
-    protected String getString(@StringRes int stringRes) {
-        return mContext.getString(stringRes);
-    }
-
-    protected float getDimension(@DimenRes int dimenRes) {
-        return mContext.getResources().getDimension(dimenRes);
-    }
-
-    protected int getDimensionPixelSize(@DimenRes int dimenRes) {
-        return mContext.getResources().getDimensionPixelSize(dimenRes);
     }
 
 }
